removed conditional CSR from _app.tsx
